requirements.txt
# AGC AI 庫存查詢系統 - 依賴套件

# Database
psycopg2-binary==2.9.9

# HTTP Client
requests==2.31.0
httpx>=0.25.0

# RAG extraction
pypdf>=4.0.0

# API Server
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3

# Environment
python-dotenv==1.0.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
